Learnt about XML serialization
I already knew about JSON (de)serialization though the `System.Text.Json` namespace introduced with .NET Core 3, but didn't know anything about XML serialization. Since the `JsonSerializer.Deserialize<T>(string)` static method doesn't have a simple counterpart for XML, I took the occasion of this test to learn a little more about XML serialization by creating a static class that would mimic (a fraction of) the API of `JsonSerializer`.

After a few hours of research, it looks like the easy way to customize the behavior of the `XmlSerializer` (e.g. use attributes instead of elements, snake_case renaming, etc...) seems to be via the application of attributes on the data-model objects' properties. I am not a fan of this method, because it clutters the data structure's source code with the concern of serialization.

After a few more hours of coding, I managed to get the behavior I wanted with the `PortersonXmlSerializer` static class.

- It certainly performs poorly, since it uses reflection.
- It is certainly incomplete regarding the cases it can handle.
- It is certainly under-tested.
- I have certainly been influenced by the fact that this is a greenfield project: existing projects probably already use attributes on objects' properties to customize the serializer's behavior; I would have followed this convension.
- Nonetheless, it was much fun.

This is how I lost my Friday.
._.