Part 4. Builder solution
One approach to part 4 could be to copy part 1, and to update the attributes applied to the data model objects properties in order to change the behavior of the XML serializer.

However, I took the occasion to get back to my original utopia of *the attributeless data structure*, and explore the path of configuring the `XmlAttributeOverrides? overrides` parameter of the `XmlSerializer`'s constructor, hence moving the concern of serialization entirely on the service side. I took inspiration from [`ExtendedXmlSerializer`](https://github.com/ExtendedXmlSerializer/home) and the syntax of its configuration DSL, and landed on a lambda-based `XmlSerializerBuilder` to customize each type:

```cs
book
    .AttributeMember(book => book.PublishDate, "publish_date", "date")
```

I could have expanded the principle by complexifying the configuration of each property, but the code would have become far too verbose for our case:

```cs
book
    .Member(book => book.PublishDate, publishDate => publishDate
        .Attribute()
        .Name("publish_date")
        .DataType("date"))
```

Certainly undertested.
._.